First commit to push google authentication with Firebase integration into the project. This a non complete submit as we need to connect it with the right Firebase account. To do so, I will contact Luca, our Firebase expert.
Therefore, this code only show a button sign in with google that when trigerred, opens the common google sign in page. However, since Firebase is not connected, the last page showing the user information cannot shows up yet.

This code was largely inspired by the bootcamp and the following video:
https://www.youtube.com/watch?v=jEKjI0OzNNI