Added view to follow and unfollow users (#86)
Note: some tests are still not passing. I discussed about it with David and he might have spot the cause of the bugs. Since the code coverage is already above 80% for this PR, I prefer to merge now and try to debug those bugs at the beginning of the next sprint.